backend/pyproject.toml
[project]
name = "photo-retriever-backend"
version = "0.1.0"
description = "Backend API for intelligent photo retrieval system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic-settings>=2.0.0",
    "structlog>=23.2.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "python-multipart>=0.0.22",
    "motor>=3.7.1",
    "beanie>=2.0.1",
    "email-validator>=2.3.0",
    "python-dotenv>=1.2.1",
    "pinecone-client[grpc]==3.2.2",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = []

